refactor: ProcessedInput, RacingReq 수정
ProcessedInput
데이터를 주고 받는 용도로 사용하기에 dto 디렉토리 아래로 위치를 옮겼습니다.
클래스명도 RacingInfoDTO로 수정해줬습니다.

RacingReq
용도를 보다 명확히 드러내기 위해 이름을 RacingInputReq로 변경했습니다.
