+++ src/main/java/racingcar/dto/RacingInfoDTO.java
@@ -1,13 +1,15 @@
-package racingcar.model;
+package racingcar.dto;
+
+import racingcar.model.Car;
 
 import java.math.BigInteger;
 import java.util.List;
 
-public class ProcessedInput {
+public class RacingInfoDTO {
     private final List<Car> carList;
     private final BigInteger raceRounds;
 
-    public ProcessedInput(List<Car> carList, BigInteger raceRounds) {
+    public RacingInfoDTO(List<Car> carList, BigInteger raceRounds) {
         this.carList = carList;
         this.raceRounds = raceRounds;
     }
